+++ lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_tab_helper.rb
@@ -9,11 +9,6 @@ module OnlyofficeWebdriverWrapper
       execute_javascript('window.open()')
     end
 
-    def maximize
-      @driver.manage.window.maximize
-      OnlyofficeLoggerHelper.log('Maximized current window')
-    end
-
     def resize_tab(width, height)
       @driver.manage.window.resize_to(width, height)
       OnlyofficeLoggerHelper.log("Resize current window to #{width}x#{height}")
@@ -60,10 +55,6 @@ module OnlyofficeWebdriverWrapper
       @driver.close
       sleep 1
       switch_to_main_tab
-    end
-
-    def close_window
-      @driver.close
     end
 
     def close_popup_and_switch_to_main_tab