+++ spec/attribute_helper_spec.rb
@@ -35,5 +35,36 @@ describe 'OnlyofficeWebdriverWrapper::WebDriver#WebdriverAttributesHelper' do
     end
   end
 
+  describe '#get_attributes_of_several_elements' do
+    it 'get_attributes_of_several_elements if elements exists' do
+      expect(webdriver.get_attributes_of_several_elements('//div', 'custom-attribute'))
+        .to eq(%w[first second])
+    end
+
+    it 'get_attributes_of_several_elements if no elements' do
+      expect(webdriver.get_attributes_of_several_elements('//div', 'no-attribute'))
+        .to eq([nil, nil])
+    end
+
+    it 'get_attributes_of_several_elements if not-existin element' do
+      expect(webdriver.get_attributes_of_several_elements('//no', 'no-attribute'))
+        .to be_empty
+    end
+  end
+
+  describe '#get_index_of_elements_with_attribute' do
+    it 'get_index_of_elements_with_attribute for several objects' do
+      expect(webdriver.get_index_of_elements_with_attribute('//span', 'custom-attribute', 'findable')).to eq(1)
+    end
+
+    it 'get_index_of_elements_with_attribute for several only_visible false' do
+      expect(webdriver.get_index_of_elements_with_attribute('//span', 'custom-attribute', 'findable', false)).to eq(2)
+    end
+
+    it 'get_index_of_elements_with_attribute for non-existing object' do
+      expect(webdriver.get_index_of_elements_with_attribute('//no', 'custom-attribute', 'findable')).to eq(0)
+    end
+  end
+
   after { webdriver.quit }
 end